fix: update ddd structure folder
