+++ src/ms-user/auth/application/dto/signin.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
-import { normalizeEmailTransform } from '../../../../../common/utils/email-normalization.util';
+import { normalizeEmailTransform } from '../../../../common/utils/email-normalization.util';
 
 export class SignInDto {
   @ApiProperty({ example: 'jhon@gmail.com', description: 'User email' })
