+++ src/ms-user/auth/application/services/singin.use-case.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UserProps } from '../../../user/domain/entities/user.entity';
 import { SecurityServiceInterface } from '../../domain/repositories/security-service.repository-interface';
-import { SignInResponseDto } from '../../infrastructure/http/dto/signin-response.dto';
+import { SignInResponseDto } from '../dto/signin-response.dto';
 
 @Injectable()
 export class SignInUseCase {
