+++ src/ms-user/auth/domain/repositories/security-service.repository-interface.ts
@@ -1,5 +1,5 @@
 import { UserProps } from '../../../user/domain/entities/user.entity';
-import { SignInDto } from '../../infrastructure/http/dto/signin.dto';
+import { SignInDto } from '../../application/dto/signin.dto';
 import { TokenPayloadDto } from '../../infrastructure/security/dto/token-payload.dto';
 
 export interface SecurityServiceInterface {
